refactor: move task helpers and utils to task level
